a2a/requirements.txt
a2a-sdk
httpx
uvicorn
python-dotenv
authlib
fastapi
starlette
itsdangerous
google-auth
requests
gTTS
python-multipart
SpeechRecognition
pydub